:beetle: corrigindo bug perfil
