+++ src/components/FormEditAddress/styled.js
@@ -1,0 +1,16 @@
+import styled from  'styled-components'
+
+export const ContainerForm = styled.div`
+    display: flex;
+    justify-content: center;
+    margin: 64px 0 80px 0;
+`
+
+export const InputsContainer = styled.div`
+    display: flex;
+    flex-direction: column;
+    width: 80vw;
+    max-width: 450px;
+    align-items: center;
+    margin-bottom: 20px;
+`
