+++ src/components/UserAdress/UserAdress.js
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { useContext } from 'react'
 import { useHistory } from 'react-router-dom'
 import useProtectedPage from '../../hooks/useProtectedPage'
 import { goToEditAdress } from '../../routes/coordinator'
@@ -6,11 +6,13 @@ import editPen from '../../assets/editPen.png'
 import { ContainerAdress } from './styled'
 import Loading from '../Loading/Loading'
 import useRequestData from '../../hooks/useRequestData'
+import { GlobalStateContext } from '../../global/GlobalStateContext'
 
 const UserAdress = () =>{
     const history = useHistory()
 
     useProtectedPage()
+    // const {address, getAddress} = useContext(GlobalStateContext)
     const address = useRequestData({}, `/profile/address`)
 
     return(
@@ -21,7 +23,7 @@ const UserAdress = () =>{
                     <p id={"adress"}>{`${address.address.street}, ${address.address.number}, ${address.address.complement} ${address.address.neighbourhood} 
                     ${address.address.city}-${address.address.state} `}</p>
                 </div>
-                    <img src={editPen} alt={"Caneta de editar"} onClick={() => goToEditAdress(history)}/>
+                    <img src={editPen} alt={"Caneta de editar"} onClick={() => {goToEditAdress(history)}}/>
             </ContainerAdress>: <Loading/>}
         </div>
     )
